feat: basing zoom/pan, requestAnimationFrame

+++ wasm-src/input/input_manager.h
@@ -112,7 +112,7 @@ private:
   bool on_key_up();
 
   bool on_resize(int x, int y, int offset_x, int offset_y);
-  bool on_wheel(PointerTarget target, int delta_x, int delta_y);
+  bool on_wheel(PointerTarget target, float delta_x, float delta_y);
 
   bool on_clipboard_copy();
   bool on_clipboard_paste();
